refactor: 단위 테스트 대상 클래스와 동일한 패키지로 분리
