+++ src/test/java/domain/strategy/CustomTicketGenerator.java
@@ -1,11 +1,11 @@
-package domain;
+package domain.strategy;
 
 import java.util.Iterator;
 import java.util.List;
 
 import domain.strategy.TicketGenerator;
 
-class CustomTicketGenerator implements TicketGenerator {
+public class CustomTicketGenerator implements TicketGenerator {
 
 	private Iterator<List<Integer>> numbers;
